fix: extract chart drawing helpers and improve test coverage
